Move all util classes in utils folder
