+++ package.json
@@ -76,7 +76,6 @@
     "typescript": "^5.6.2"
   },
   "dependencies": {
-    "@redhat-developer/vscode-redhat-telemetry": "^0.9.0",
-    "extract-zip": "^2.0.1"
+    "@redhat-developer/vscode-redhat-telemetry": "^0.9.0"
   }
 }
